% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd.R
\name{lsd}
\alias{lsd}
\title{Diseno de un experimento alpha lattice}
\usage{
lsd(x, y)
}
\arguments{
\item{x}{Vector columna}

\item{y}{Vector columna}
}
\value{
Diseno del experimento
}
\description{
Diseno de un experimento usando Alpha Lattice (diseño utilizado en mejoramiento genetico de trigo).
}
